tms_core/requirements.txt
djangorestframework-simplejwt>=5.0
Django>=5.0,<6.0
psycopg2-binary>=2.9
Pillow>=10.0
djangorestframework>=3.14
django-cors-headers>=4.3
gunicorn>=21.2
redis>=5.0
celery>=5.3
django-celery-beat>=2.5
channels>=4.0
daphne>=4.0
requests
